add endpoints to update user details and retrieve user by ID

+++ identity-service/src/main/java/com/javatechie/repository/UserCredentialRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface UserCredentialRepository  extends JpaRepository<UserCredential,Integer> {
+public interface UserCredentialRepository  extends JpaRepository<UserCredential,Long> {
     Optional<UserCredential> findByEmail(String email);
     Optional<UserCredential> findByResetToken(String token);
